+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - "You miss 100 percent of the chances you don't take. — Wayne Gretzky" — Michael Scott
 
 - [apm] fix: Use Performance API for timings when available, including Web Workers (#2492)
+- [apm] fix: Remove Performance references (#2495)
 
 ## 5.14.1